button arrays and status message improvements
You can now generate a set of toggle buttons from a lookup, to use as criteria for further lookups. The status message now uses an overlay to be more prominent, and the 'no data' messaging is now persistent - add a "nodata":"mydiv" clause to the config JSON, and add a suitable DIV element containing an appropriate message under the 'tab' element with a class of 'hidden'.