Fixed fullcalendar issues
When writing a new event, the database could return a new event id, but fullcalendar would ignore this. This meant that subsequent changes to a new event would not save properly. The code will now save the new event's ID as an extended property called sql_id, and then when making subsequent changes it will pass this as the eventID if the eventID is blank.